2021/01

+++ createAOC.js
@@ -8,14 +8,18 @@ const content = `const fs = require("fs");
 const args = process.argv.splice(2);
 const fileName = args[0] === "demo" ? "./demo" : "./input";
 
+
 let input = fs
   .readFileSync(fileName, "utf-8")
   .split(/\\n/)`;
 
 const createDir = (dirPath) => fs.mkdirSync(dirPath);
 
-const createFile = (filePath, fileContent) =>
-  fs.writeFileSync(filePath, fileContent);
+const createFile = (filePath, fileContent) => {
+  if (!fs.existsSync(filePath)) {
+    fs.writeFileSync(filePath, fileContent);
+  }
+};
 
 function createAoCDay(year, day) {
   let path = `./${year}/Day\ ${day}`;
@@ -26,8 +30,8 @@ function createAoCDay(year, day) {
 }
 
 function createDay(year, day) {
-  if(!fs.existsSync(`./${year}/Day\ ${day}`)) {
-    createAoCDay(year,day)
+  if (!fs.existsSync(`./${year}/Day\ ${day}`)) {
+    createAoCDay(year, day);
   }
 }
 
@@ -36,11 +40,14 @@ function createYear(year) {
     createDir(`./${year}`);
   }
 
-  for (let i = 1; i <= 12; i++) {
-    let day = i < 10 ? `0${i}` : i;
-    if(!fs.existsSync(`./${year}/Day\ ${day}`)) {
-      createAoCDay(year,day)
+  for (let i = 1; i <= 24; i++) {
+    let day = i < 10 ? `0${i}` : i.toString();
+    let currDate = `./${year}/Day\ ${day}`;
+    if (!fs.existsSync(currDate)) {
+      console.log(currDate);
+      createAoCDay(year, day);
     }
   }
 }
-createYear(year)
+
+createYear(year);
